+++ src/nvlib/model/ods/ods_w_sectionlist.py
@@ -4,9 +4,9 @@ Copyright (c) 2025 Peter Triesberger
 For further information see https://github.com/peter88213/novelibre
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
-from nvlib.model.data.section import Section
 from nvlib.model.ods.ods_w_grid import OdsWGrid
 from nvlib.novx_globals import SECTIONLIST_SUFFIX
+from nvlib.novx_globals import STATUS
 from nvlib.nv_locale import _
 
 
@@ -367,5 +367,5 @@ class OdsWSectionList(OdsWGrid):
             sectionNumber,
             wordsTotal
         )
-        sectionMapping['Status'] = Section.STATUS[sectionMapping['Status']]
+        sectionMapping['Status'] = STATUS[sectionMapping['Status']]
         return sectionMapping